requirements.txt
streamlit
openai
anthropic
google-generativeai
python-dotenv
PyPDF2
google-ai-generativelanguage
google-api-core